requirements.txt
pandas>=2.2
openpyxl>=3.1
httpx[http2]>=0.28
boto3>=1.34
sentence-transformers>=0.6
scikit-learn>=1.4
tqdm>=4.66
PyYAML>=6.0
